Refactor app into 1 component per file.
