+++ src/browser/WebViews.js
@@ -1,0 +1,12 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+'use strict';
+
+const {List} = require('typed-immutable/index');
+const WebView = require('./WebView');
+
+const WebViews = List(WebView);
+
+module.exports = WebViews;
