+++ src/browser/theme.js
@@ -26,7 +26,7 @@ const createTheme = (properties) => Object.assign({
   navigationPanel: {backgroundColor: '#fff'},
   progressbar: {color: '#82D3FD'},
   awesomebarSuggestions: {backgroundColor: '#fff', color: 'rgba(0,0,0, 0.65)'},
-  gossamerButton: {color: 'green'}
+  gossamerButton: {color: '#444'}
 }, properties);
 
 const IS_DARK = true;
